requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
